Implement a facts segment
The FactsSegment class allows both scanning the facts segment and using the b-tree to find a start position suitable to a given start condition. Note that the facts segment does not understand the value ordering, i.e., it cannot map value1/value2/value3 into subject/predicate/object. This should be done by higher-level operators, as the ordering is different for each facts segment.