implement skipping during merge joins
The IndexScan and AggregatedIndexScan operators now allow for giving scan hints.
Whenever the operator reaches a new page, it checks the hint and discards the
whole page if the next plausible value is behind this page. When skipping a
page, the B-tree is used to find the next valid page.
The hints are only checked at the beginning of a page as the checks themselves
are relatively expensive.