implement buffering in TempFile
In theory this change is unncessary, and might be reverted in the future. In
practice, this avoids a strange output file corruption for very long records.
Perhaps the ostream implementation of g++ does not like mixing put and write,
or it is a strange memory corruption bug. Will investigage further.