use a lookup table for counts instead of constantly scanning
The statistics builder checks the number of occurances of a value extremely
often. When using the FullyAggregatedIndex for this, the index is scanned
and re-scanned constantly, where most time is actually spend on decompressing
pages. Instead, precompute a simple map with a static layout and map it
into the address space for lookups.